(REFACTOR): created directory config
