+++ ExternalService/src/main/java/org/example/externalservice/config/WebConfig.java
@@ -1,9 +1,9 @@
-package org.example.externalservice;
+package org.example.externalservice.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.web.config.EnableSpringDataWebSupport;
 
 @Configuration
 @EnableSpringDataWebSupport(pageSerializationMode = EnableSpringDataWebSupport.PageSerializationMode.VIA_DTO)
-public class WebConfig {
+public class  WebConfig {
 }
